Create pathtype.Path and the `exists` validation
The pathtype.Path class can be passed as `type` parameter to
`argparse` arguments. It will automatically convert the argument to a
`pathlib.Path` instance.

It also supports custom validation callables through it's `validator`
parameter. If any of those validator raises an exception, it will be
caught by `argparse` which will show a nice error message.

The class comes with one predefined validation: `exists` which validates
that the path points to an existing file or directory.

Create the `pathtype.Path`, the `exists` validation and all associated
tests.