Add read, write and execute predefined validations
The user might want to validate that they have read, write or execute
access on the file or directory at the path.

Create new predefined validations for those 3 types of check and add
new triggers in the Path() class to add them as validators. Create new
tests to test the new validations.