Add a "doesn't exist" predefined validation
The user might want to validate that the path doesn't point to an
existing file or directory. Or if it's a symlink, that it doesn't point
to an existing file or directory.

Create a new predefined validation to check for non-existence and add a
new trigger in the Path() class to automatically add the validator.
Create tests to test the new validation.