Add a "parent directory exists" predefined validation
When receiving a file path as command line argument, we might accept
that no file exists at that path, but we may require that its parent
directory exists. For example, we ask the user the file path where to
output results, and we want to make sure that the directory already
exists.

Create a new predefined validation to check for the existence of the
parent directory, and add a new trigger in the Path() class to
automatically add the validator. Create tests to test the new
validation.