Add a "parent directory writable" predefined validation
When receiving a file path as command line argument, we might want to
check if it's possible to create the file or directory. A requirement to
create the file is that the user has "write" permissions on the parent
directory of the path.

Create a new predefined validation to check that the user has "write"
permission on the parent directory of the path. Create tests.