Add an "All" predefined validation
We want to build more complex validations by creating a conjunction of
validators (an "AND" list of validators). If any validator fails, the
whole validation fails, else it succeeds.

Create an `Any` validation class that accepts a list of validators and
that fails as soon as any children fails. Create tests.