Add a "path creatable" trigger in Path
When parsing a path in the arguments, we might want to validate that the
user has the permissions to create this path in its parent directory.
Specifically, we would want to validate that the parent directory exists
and that the user has "write" permissions on it.

Create a new single trigger in the `Path` class that checks if 1) the
parent directory exists and 2) that the current user has "write"
permission on it. Create a test.