Add a "path creatable or writable" trigger in Path
A common use case is when we want to make sure we can write to a file,
or, if it doesn't exist yet, that we can create it. Think of a log file
that we want to create if it doesn't yet exist, to then append to it.
So if the file doesn't exist, act like if ``creatable`` is True, else
act like ``writable`` is True.

Create a new trigger in the `Path` class that acts like if ``creatable``
is True if the path doesn't exist, else acts like ``writable`` is True.
Create a test.