Add a "name matches" predefined validation
We want a predefined validation that validates if the name part of a
path (i.e. the file name, including the extension) matches either a
regular expression or a glob pattern. We also want it as one of the
"fast" validations available in the Path() class.

Create a new validation that validates that the name matches a pattern.
Also add a new argument to Path to quickly and simply use this new
validation. Create tests.