Remove argparse stderr output when running tests
Some tests assert that the parser fails under some conditions. When the
parser fails, it outputs an error message to stderr. So even if the test
passes, an error message is shown. At minimum this pollutes the test
log, at worst it can incorrectly lead to believe that the tests didn't
pass.

Update the tests validating the failure of the parser so that the error
message is not shown needlessly.