Refactor tests to better support different platforms.
Some tests specifically test when a file has specific access permissions, or
when it's a symbolic link. Those tests don't apply to Windows so they should be
skipped on that platform. Also, duplicate code was starting to appear.

Refactor tests to mark tests that use file access or symbolic links (those tests
are to be skipped on Windows). Also refactor code to remove duplicate code.
Improve test documentations.