NameMatches and PathMatches accept keyword only args
NameMatches and PathMatches __init__ method accepts positional arguments which
are the regex pattern and the glob pattern. The user has to use one OR the
other. They have the exact same role, are exclusive and one is not more
important than the other. So it's strange that if we supply a regex, we can do
it with NameMatches(regex), but if it's a glob, it has to be
NameMatches(glob=glob) (or worse: NameMatches(None, glob=glob)).

Change the __init__ signature so that both argument can only be specified as
keyword arguments.